server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "3.7"
psycopg2-binary = "^2.9.5"
cmake = "3.25.0"
django-extensions = "3.2.1"
req2toml = "^1.2.0"
asgiref = "3.6.0"
sqlparse = "0.4.3"
django = ">=3.2.4,<3.3"
pillow = "^9.4.0"
django-filter = "^22.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
